Comments be (#33)
* Adding video metadata be (#19)

* everything works except some tests need to be added and video model is incomplete

* tests doesnt work

* everything works

* solution clean up

* small changes

* Selenium + login (#20)

* Added login/out

* login finished

* started playwright

* added data seed

* fff

* Remote database (#21)

* Added Remote Database script

* Added change to appsettings

* Add video frontend (#22)

* adding the videosmetadata without service

* uploading video should work

* di fix

* gitignore ignores ffmpeg related files

* api tests added

* tests work

* added two last tests

* added a way to add metadata

* video upload frontend

* thumbnail is now required

* small changes follow specification

* get video tests work

* comments resolved

* added models for comments

* migration added

* added tests

* get all comments functionality without tests

* get all comments with tests

* add comments tests fixed

* post comment response with tests

* cascade delete of video and comment config and user to author rename

* remove comment with tests

* get all comment responses

* delete comment response without test

* added getting the user id from claims

* tests not finished

* everything done

* fixed errors after merge

* fixed some issues

* removed using

* comments resolved

* removed all unnecessary usings in all solution

* IdentityControllerBase is not abstract

---------

+++ backend/YouTubeV2/Controllers/SubscriptionsController.cs
@@ -1,18 +1,16 @@
 using Microsoft.AspNetCore.Mvc;
 using System.ComponentModel.DataAnnotations;
-using System.Security.Claims;
 using YouTubeV2.Api.Attributes;
 using YouTubeV2.Application.DTO.SubscriptionDTOS;
-using YouTubeV2.Application.Exceptions;
 using YouTubeV2.Application.Model;
 using YouTubeV2.Application.Services;
-using YouTubeV2.Application.Services.JwtFeatures;
 
 namespace YouTubeV2.Api.Controllers
 {
     [ApiController]
     [Route("subscriptions")]
-    public class SubscriptionsController : ControllerBase
+    [Roles(Role.Simple, Role.Creator, Role.Administrator)]
+    public class SubscriptionsController : IdentityControllerBase
 	{
 		private readonly ISubscriptionService _subscriptionsService;
 
@@ -21,29 +19,28 @@ namespace YouTubeV2.Api.Controllers
 			_subscriptionsService = subscriptionsService;
 		}
 
-        [Roles(Role.Simple, Role.Creator, Role.Administrator)]
         [HttpGet]
 		public async Task<ActionResult<UserSubscriptionListDto>> GetSubscriptionsAsync([FromQuery][Required] Guid id, CancellationToken cancellationToken) =>
             Ok(await _subscriptionsService.GetSubscriptionsAsync(id.ToString(), cancellationToken));
 
-        [Roles(Role.Simple, Role.Creator, Role.Administrator)]
         [HttpPost]
         public async Task<IActionResult> PostSubscriptionsAsync([FromQuery][Required] Guid subId, CancellationToken cancellationToken)
         {
-            string subscriberId = GetUserId();
+            string? subscriberId = GetUserId();
+            if (subscriberId is null) return Forbid();
+
             await _subscriptionsService.AddSubscriptionAsync(subId.ToString(), subscriberId, cancellationToken);
             return Ok();
         }
 
-        [Roles(Role.Simple, Role.Creator, Role.Administrator)]
         [HttpDelete]
         public async Task<IActionResult> DeleteSubscriptionsAsync([FromQuery][Required] Guid subId, CancellationToken cancellationToken)
         {
-            string subscriberId = GetUserId();
+            string? subscriberId = GetUserId();
+            if (subscriberId is null) return Forbid();
+
             await _subscriptionsService.DeleteSubscriptionAsync(subId.ToString(), subscriberId, cancellationToken);
             return Ok();
         }
-
-        private string GetUserId() => User.Claims.First(claim => claim.Type == ClaimTypes.NameIdentifier).Value;
     }
 }
